_cite/requirements.txt
manubot==0.5.5
PyYAML==6.0
diskcache==5.4.0
rich==12.6.0
python-dotenv==0.21.0
google-search-results==2.4.1
